feat: System sampler for macos
