+++ pkg/plugins/plugins_darwin.go
@@ -4,6 +4,9 @@ package plugins
 
 import (
 	"github.com/newrelic/infrastructure-agent/internal/agent"
+	"github.com/newrelic/infrastructure-agent/pkg/metrics"
+	metricsSender "github.com/newrelic/infrastructure-agent/pkg/metrics/sender"
+	"github.com/newrelic/infrastructure-agent/pkg/metrics/storage"
 	"github.com/newrelic/infrastructure-agent/pkg/plugins/ids"
 	"github.com/newrelic/infrastructure-agent/pkg/plugins/proxy"
 )
@@ -21,6 +24,21 @@ func RegisterPlugins(a *agent.Agent) error {
 	if config.FilesConfigOn {
 		a.RegisterPlugin(NewConfigFilePlugin(*ids.NewPluginID("files", "config"), a.Context))
 	}
+
+	sender := metricsSender.NewSender(a.Context)
+	//procSampler := process.NewProcessSampler(a.Context)
+	storageSampler := storage.NewSampler(a.Context)
+	//nfsSampler := nfs.NewSampler(a.Context)
+	//networkSampler := network.NewNetworkSampler(a.Context)
+	systemSampler := metrics.NewSystemSampler(a.Context, storageSampler)
+
+	sender.RegisterSampler(systemSampler)
+	//sender.RegisterSampler(storageSampler)
+	//sender.RegisterSampler(nfsSampler)
+	//sender.RegisterSampler(networkSampler)
+	//sender.RegisterSampler(procSampler)
+
+	a.RegisterMetricsSender(sender)
 
 	return nil
 }
